Banking app/server/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-marshmallow==1.3.0
Flask-Mail==0.10.0
python-dotenv==1.2.1
sentry-sdk==2.48.0
Flask-JWT-Extended==4.7.1
flask-cors==6.0.2
